International database with facilities to specifie exhange markets and new indicators from zipline, portfolio table, cleaning shiny, some cool tools, moved network configuration in config dir, new backtest initialization

+++ backtester/optimization/genetic_function.py
@@ -57,7 +57,7 @@ class Metric(object):
                 ordered parameters to optimize
         _________________________________________
         Return
-            score: float(1)
+            score: float
                 Error of the cost function ran with this solution
         '''
